# Batched & parallel support for cfn-lint, eslint, gitleaks