--- a/lib/functions/experimental-batch-workers/cfn-lint.sh
+++ b/lib/functions/experimental-batch-workers/cfn-lint.sh
@@ -15,6 +15,36 @@ function LintCodebaseCfnLint() {
   info "Running EXPERIMENTAL batched LintCodebase on ${#FILE_ARRAY[@]} files. FILE_TYPE: ${FILE_TYPE}, LINTER_NAME: ${LINTER_NAME}, LINTER_COMMAND: ${LINTER_COMMAND}, TEST_CASE_RUN: ${TEST_CASE_RUN}"
   # * space separated $LINTER_COMMAND, expect to split before feeding into xargs correctly
   # shellcheck disable=SC2046 disable=SC2086
+  CFNLINT_ERRORS_FOUND=$(
+    printf "%s\n" "${FILE_ARRAY[@]}" | \
+    parallel -L 16 -P "$(nproc)" --xargs ${LINTER_COMMAND} | \
+    tee "/dev/stderr" | \
+    (
+      CFNLINT_ERRORS_FOUND=0
+      while IFS= read -r LINE; do
+        if grep "[EW][0-9]\+[[:space:]]" <<<"$LINE" >/dev/null; then
+          debug "cfnlint error matched"
+          CFNLINT_IS_ERROR="true"
+          continue
+        fi
+        if grep "$PWD" <<<"$LINE" >/dev/null; then
+          CFNLINT_NEXT_FILENAME=$(cut -d: -f1 <<<"$LINE")
+          if [[ "$CFNLINT_NEXT_FILENAME" != "$CFNLINT_CUR_FILENAME" ]]; then
+            debug "Filename from cfnlint: $CFNLINT_NEXT_FILENAME"
+            CFNLINT_CUR_FILENAME=$CFNLINT_NEXT_FILENAME
+            if [[ "$CFNLINT_IS_ERROR" == "true" ]]; then
+              CFNLINT_IS_ERROR="false"
+              (("CFNLINT_ERRORS_FOUND++"))
+              debug "Error counter incremented to ${CFNLINT_ERRORS_FOUND}"
+            fi
+          fi
+          continue
+        fi
+      done
+      echo "${CFNLINT_ERRORS_FOUND}"
+    )
+  ) 2>&1
+
   # * workaround shfmt error for =ESLINT_ERRORS_FOUND
   for i in $(seq "$CFNLINT_ERRORS_FOUND");
   do