<!-- Please ensure your PR title is brief and descriptive for a good changelog entry -->
<!-- Link to issue if there is one -->
<!-- markdownlint-disable -->

Proposal to fix
- #4084  

Currently planned to provide implementation for
- [x] cfn-lint
- [x] eslint
- [x] gitleak

Use cases includes
- Faster whole repo linting on cloud (changed files can affect unchanged files)
- Local linting on commit / push without having developers to wait too much

<!-- markdownlint-restore -->

<!-- Describe what the changes are -->

## Proposed Changes

_Batched_ and _parallel_ linting for use case involving large number of files.
- _Batched_: Use `$LINTER_COMMAND FILE1 FILE2 FILE3 ....` instead of `$LINTER_COMMAND FILE` looping over all files
- _Parallel_: Files are divided into not-so-small chunks and multiple linter instances up to `nprocs` are run in parallel
As for most parallelization, this patch does not reproduce exact output of serial version

To enable this experimental worker, set env `EXPERIMENTAL_BATCH_WORKER` to `true`

In order to mimic serial behavior as closely as possible, the implementation
- show linter output as is, 
- does not break or interleave linter output in any way,
- provide the same error file count (lines outputting text such as `[ERROR]   ERRORS FOUND in JAVASCRIPT_ES:[123]`) as serial version when super-linter is completed

Current ERROR output on serial version cannot be recovered, since
- it is not possible retrieve error code of linter for individual file, batching implies only 1 return code is given for a batch of files.

### Performance comparison

A figure is provided only for eslint on a repo that I have on hand
- 8 core laptop
- approx. 450 .js file
- timing only on JAVASCRIPT_ES block

Test 1: 429 files contains linter error (not warning) artificially introduced (mainly `react/*` rules that was turned off for the project)
- original: `7 min 21 sec`
- new impl: `14 sec` (it is not necessary to parallel, still under 30s without parallel)

Test 2: No artificially introduced error (0 error, several warnings)
- original: `7 min 4 sec`
- new impl: `5 sec`

For the experimental implementation, majority of the performance issue is in bash-script implemented output parser to calculate number of error files, it is not too bad anyway.

## Readiness Checklist

### Author/Contributor
- [x] If documentation is needed for this change, has that been included in this pull request
- [ ] Test case count is not yet recovered

### Reviewing Maintainer
- [x] Label as `breaking` if this is a large fundamental change
- [x] Label as either `automation`, `bug`, `documentation`, `enhancement`, `infrastructure`, or `performance`