@kftsehk 2023-04-17
Just wonder why this gets a linter error, was assigning the internal counter back to variable `ERRORS_FOUND_${FILE_TYPE}`
```
  (("ERRORS_FOUND_${FILE_TYPE}"=CFNLINT_ERRORS_FOUND));
```