@kftsehk 2023-04-17
I think I will leave the python toolchain later, but adding gitleaks parallelized (gitleaks does not support batch, only parallel is possible)

One small note for gitleaks

parallel version only output the information related to the detected leaks
parallel version does not output
- (number|"no") leaks found
- gitleaks took (number) ms
because when parallel, it is not possible to relate the above message to the file which generates it.

The number returned at the end of super-linter is still correct, equals to serial run.