@kftsehk 2023-04-17
Sorry for messing with linter errors, have checked locally against this for latest commit

```
docker run --rm  -e IGNORE_GITIGNORED_FILES=true        -e LINTER_RULES_PATH=/     -e LOG_LEVEL=NOTICE     -e RUN_LOCAL=true     -e USE_FIND_ALGORITHM=true     -e VALIDATE_BASH=true     -e ERROR_ON_MISSING_EXEC_BIT=true -e VALIDATE_BASH_EXEC=true  -e VALIDATE_SHELL_SHFMT=true  -w /tmp/lint     -v "$(git rev-parse --show-toplevel)":/tmp/lint github/super-linter:slim-latest
```