@kftsehk 2023-04-28
Hi, would like to follow up on how to proceed with testing and moving towards merging the PR.

This issue features a new code path, that is slightly incompatible (as in exact char-by-char output diff sense), enabled by setting EXPERIMENTAL_BATCH_WORKER to true, it seem current CI won’t cover this path unless also test with this env set.