@zkoppert 2023-10-25
Wow! Those are some huge performance improvements. Thanks for doing all the work on this experimental feature and getting the docs updated too. Can you take a look at resolving the file conflicts and I can get this one in the queue to merge through?